Add core authentication code and responder utility class
Adds most core functionality for the authentication of requests.
All thas is missing is the logic to generate the string used for
signing the request based on the request headers and VanGoH struct
configuration.

Additionally added a utility class, "responder" to handle formatting
and responding with appropriate HTTP responses under different
failure conditions.